Add files via upload
Uploaded results from training on shaped reward with DDPG and actor + critic NN's with two 64 dense layers. The joints of the knees multiplied by a coefficient were added to the reward. The formula was as follows: delta_x - (math.sqrt(lig_pen) * 10e-8) + -c1*knee_r_angle + -c2*knee_l_angle) where c1 and c2 are two constants. The number for the knee angles ranged from -2.0943951 to 0.17453293. For example, the rewardShape5_5 was run for 10,000 steps wth c1=.5, c2=.5. rewardShape005_005_100000 was run for 100,000 steps with c1=.005, c2=.005.